Made A Simple User Class
Made a simple user class that just has a logged in boolean. For the default constructor
it sets the logged in value to be false by default. There is an Overridden constructor as
well to set the value at initialization. It has 100% test coverage. And no checkstyle errors.
When we implement the account information we will update the logged in boolean to be true.
And when we implement the sign out we will make set the value to false;