Created the Freelancer
Created the Freelancer class to extend the User, adding a Categories Enum and a skills array of size 5.
Added constructors to initialize freelancers with required atributes, including the option to set skills
at creation. Implemented methods for setting and replacing skills at creation. Implemented methods for
setting and replacing skills, validating input, and checking if the freelancer contains a specific skill.
Developed a test class with 100 percent test coverage. And no checkstyle errors.