Sorting Freelancers by Categories
I made a FreelancerSorter class that has a static method that accepts a list and modifys the list
to be in the order based on their category in alphabetical order. If the list is passed in as null
it will throw an IllegalArgumentException to prevent error. I tested the sorter by testing a null
list, empty list, single freelancer list, multiple freelancer list, and presorted list, along with
all the variations of the multiple freelancer list. There is 100 percent code coverage and no
checkstyle errors.