Attempting to implment the User to Freelancer
Added some inof to the ccount page and noticed that we have no way of switching from a user to a freelancer, so attempting to implment that. Currently everything is broken and I have been working to long so I am getting off for tonight and will try to pick it up again once I recharge.