Implmented the switching between user/freelancer
A user can now switch to a freelancer if they selected the button within account to do so. Next I need to clean the account page hiding buttons as needed and need to make sure when a user switch to a new category, they can be found in the categories category on the category page.