Made the UI for Selecting skills and category
Made the UI for selecting skills. A combo box will be clicked to add the skills to
the list view. all the skills in the list view are what you have selected to search
by. When you click the list view this will be how you remove the skills from the search.
A comboBox for the category selected will let you select which category you would like
to view. The searching by name still works and is functional.